// src/app/shared/models/Funcionario.ts
export class Funcionario {
    CODIGO: number;
    CHAPA: string;
    NOME: string;
    RUA: string;
    NUMERO: string;
    BAIRRO: string;
    CIDADE: string;
    ESTADO: string;
    SEXO: string;
    CEP: string;
    CPF: string;
    DIGCPF: string;
    CARTIDENTIDADE: string;
    SITUACAO: string;
    TIPO: string;
    SALARIO: number;
    COLIGADA: string;
    FILIAL: string;
    AGENCIA: number;
    BANCO: string;
    DIGAG: number;
    CONTACORRENTE: number;
    DIGCONTA: number;
    DATAADMISSAO: Date;
    DTAVEMCFERIAS: Date;
    CODSECAO: string;
    CODCOLIGADA: number;
    CODFILIAL:number;
    FUNCAO: string;
    CODEQUIPE: number;
    FUNCAO_COD: number;
    DTA_ALTER: Date;
    TELEFONE1: string;
    TELEFONE2: string;
    TELEFONE3: string;
    AJUDACUSTO: Date;
    NOMEFILIAL: string;
    FANTASIAFILIAL: string;
    SESSAO: string;
    SEQPESSOA: number;
    UNIDADE: string;

}

export class FuncGrid {
    unidade: string;
    nome: string;
	chapa: string;
	cargo: string;
	depto: string;
	empresa: string;
    usercomp: string;
    situacao: string;
}